Replace forge error screen with new implementation from architectury api.
